# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TM Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/shots:
    get:
      operationId: listShots
      tags: [shots]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of shots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shot"
    post:
      operationId: createShot
      tags: [shots]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShotInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shot"
  /shots/{id}:
    patch:
      operationId: updateShot
      tags: [shots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShotUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shot"
    delete:
      operationId: deleteShot
      tags: [shots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Call Sheets
...
  /projects/{projectId}/staging-environments:
    get:
      operationId: listStagingEnvironments
      tags: [staging-environments]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StagingEnvironment"
    post:
      operationId: createStagingEnvironment
      tags: [staging-environments]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StagingEnvironmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StagingEnvironment"
  /staging-environments/{id}:
    patch:
      operationId: updateStagingEnvironment
      tags: [staging-environments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StagingEnvironmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StagingEnvironment"
    delete:
      operationId: deleteStagingEnvironment
      tags: [staging-environments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Project Resources
...
components:
  schemas:
...
    Shot:
      type: object
      required: [id, projectId, userId, shotNumber, shotType, sortOrder, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        userId: { type: string }
        shotNumber: { type: string }
        description: { type: ["string", "null"] }
        shotType: { type: string }
        angle: { type: ["string", "null"] }
        lens: { type: ["string", "null"] }
        movement: { type: ["string", "null"] }
        locationNotes: { type: ["string", "null"] }
        storyboardUrl: { type: ["string", "null"] }
        sortOrder: { type: integer }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    ShotInput:
      type: object
      required: [shotNumber]
      properties:
        shotNumber: { type: string }
        description: { type: string }
        shotType: { type: string }
        angle: { type: string }
        lens: { type: string }
        movement: { type: string }
        locationNotes: { type: string }
        storyboardUrl: { type: string }
        sortOrder: { type: integer }
        status: { type: string }
    ShotUpdate:
      type: object
      properties:
        shotNumber: { type: string }
        description: { type: string }
        shotType: { type: string }
        angle: { type: string }
        lens: { type: string }
        movement: { type: string }
        locationNotes: { type: string }
        storyboardUrl: { type: string }
        sortOrder: { type: integer }
        status: { type: string }
...
    StagingEnvironment:
      type: object
      required: [id, projectId, userId, name, environment, branch, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        userId: { type: string }
        name: { type: string }
        environment: { type: string }
        url: { type: ["string", "null"] }
        branch: { type: string }
        status: { type: string }
        notes: { type: ["string", "null"] }
        lastDeploy: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    StagingEnvironmentInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        environment: { type: string }
        url: { type: string }
        branch: { type: string }
        status: { type: string }
        notes: { type: string }
    StagingEnvironmentUpdate:
      type: object
      properties:
        name: { type: string }
        environment: { type: string }
        url: { type: string }
        branch: { type: string }
        status: { type: string }
        notes: { type: string }
        lastDeploy: { type: string }
...