# lib/api-zod/src/generated/types/shotUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TM Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ShotUpdate {
  shotNumber?: string;
  description?: string;
  shotType?: string;
  angle?: string;
  lens?: string;
  movement?: string;
  locationNotes?: string;
  storyboardUrl?: string;
  sortOrder?: number;
  status?: string;
}